feat: "preview" deployments (#101)
This adds a `--preview <name>` flag to deploy/delete commands, as well as to all the env commands. This enables you to do "preview" deployments, that follow the pattern `<preview name>.<project name>.<user name>.partykit.dev`.  We'll soon build a github action that can automate this for github PRs, but feel free to use this command yourself for now.